CI: Precompile packages to prevent timeouts
When the linter runs it first needs to compile the packages, this can
cause timeouts on the linter.
We could bump the timeout on the linter, but instead we can just
precompile the packages so the linter doesn't have to, which takes
compilation out of the equation for the linter timeout.

(cherry picked from commit 80564266f6d950d6860fa197b49ec7546645b97a)
Signed-off-by: Brian Goff <cpuguy83@gmail.com>